+++ alb.tf
@@ -1,3 +1,15 @@
+# Reference the existing ACM certificate by its ARN
+#data "aws_acm_certificate" "gitea_alb_cert" {
+# domain = "gita-demo.emiring.net"
+#  arn = "arn:aws:acm:ap-southeast-1:503561451477:certificate/3fa28755-6e2f-409d-bc87-c568b58aeb42"
+#}
+
+data "aws_acm_certificate" "gitea_amz_issued" {
+  domain      = "gitea-demo.emiring.net"
+  types       = ["AMAZON_ISSUED"]
+  most_recent = true
+}
+
 # Create the Application Load Balancer
 resource "aws_lb" "cog_alb" {
   name               = "cog-alb"
@@ -29,13 +41,31 @@ resource "aws_lb_target_group" "cog_tg" {
 }
 
 # Create a listener for the ALB
-resource "aws_lb_listener" "cog_listener" {
+resource "aws_lb_listener" "cog_listener_https" {
+  load_balancer_arn = aws_lb.cog_alb.arn
+  port              = 443
+  protocol          = "HTTPS"
+#  ssl_policy        = "ELBSecurityPolicy-2016-08" #Default Security Policy
+  ssl_policy        = "ELBSecurityPolicy-TLS13-1-2-2021-06"
+  certificate_arn   = data.aws_acm_certificate.gitea_amz_issued.arn
+  default_action {
+    type             = "forward"
+    target_group_arn = aws_lb_target_group.cog_tg.arn
+  }
+}
+
+#Redirect HTTP listener with a redirect to HTTPS
+resource "aws_lb_listener" "cog_listener_http" {
   load_balancer_arn = aws_lb.cog_alb.arn
   port              = 80
   protocol          = "HTTP"
   default_action {
-    type             = "forward"
-    target_group_arn = aws_lb_target_group.cog_tg.arn
+    type = "redirect"
+    redirect {
+      protocol = "HTTPS"
+      port     = "443"
+      status_code = "HTTP_301"
+    }
   }
 }